feat(quota): ✨ add GitHub Copilot quota display and improve precision
- Implement CopilotQuotaFetcher with GitHub API integration
- Add Copilot and Kiro provider icons and enum support
- Add CLI auth commands for Copilot/Kiro login flows
- Change quota percentage from Int to Double for precision
- Fix plan display mapping ('individual' → 'Pro')
- Add quotaLookupKey to resolve Copilot account key mismatch
- Add refresh buttons to Quota screen
- Migrate to CLIProxyAPIPlus repository
- Simplify QuotaFetchError enum (remove unused cases)

+++ Quotio/Services/OpenAIQuotaFetcher.swift
@@ -190,12 +190,12 @@ struct CodexQuotaData: Codable, Sendable {
         }
     }
     
-    var sessionRemainingPercent: Int {
-        100 - sessionUsedPercent
+    var sessionRemainingPercent: Double {
+        Double(100 - sessionUsedPercent)
     }
     
-    var weeklyRemainingPercent: Int {
-        100 - weeklyUsedPercent
+    var weeklyRemainingPercent: Double {
+        Double(100 - weeklyUsedPercent)
     }
     
     func toProviderQuotaData() -> ProviderQuotaData {
